+++ packages/monaco/src/browser/monaco-editor-model.ts
@@ -20,11 +20,11 @@ import { DisposableCollection, Disposable } from '@theia/core/lib/common/disposa
 import { Emitter, Event } from '@theia/core/lib/common/event';
 import { CancellationTokenSource, CancellationToken } from '@theia/core/lib/common/cancellation';
 import { Resource, ResourceError, ResourceVersion } from '@theia/core/lib/common/resource';
-import { Saveable, SaveOptions } from '@theia/core/lib/browser/saveable';
+import { Saveable, SaveOptions, SaveReason } from '@theia/core/lib/browser/saveable';
 import { MonacoToProtocolConverter } from './monaco-to-protocol-converter';
 import { ProtocolToMonacoConverter } from './protocol-to-monaco-converter';
 import { ILogger, Loggable, Log } from '@theia/core/lib/common/logger';
-import { IIdentifiedSingleEditOperation, ITextBufferFactory, ITextModel, ITextSnapshot } from '@theia/monaco-editor-core/esm/vs/editor/common/model';
+import { ITextBufferFactory, ITextModel, ITextSnapshot } from '@theia/monaco-editor-core/esm/vs/editor/common/model';
 import { IResolvedTextEditorModel } from '@theia/monaco-editor-core/esm/vs/editor/common/services/resolverService';
 import * as monaco from '@theia/monaco-editor-core';
 import { StandaloneServices } from '@theia/monaco-editor-core/esm/vs/editor/standalone/browser/standaloneServices';
@@ -39,12 +39,7 @@ export {
     TextDocumentSaveReason
 };
 
-export interface WillSaveMonacoModelEvent {
-    readonly model: MonacoEditorModel;
-    readonly reason: TextDocumentSaveReason;
-    readonly options?: SaveOptions;
-    waitUntil(thenable: Thenable<IIdentifiedSingleEditOperation[]>): void;
-}
+export type WillSaveMonacoModelListener = (model: MonacoEditorModel, token: CancellationToken, options?: SaveOptions) => Promise<void>;
 
 export interface MonacoModelContentChangedEvent {
     readonly model: MonacoEditorModel;
@@ -83,9 +78,6 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
     protected readonly onDidSaveModelEmitter = new Emitter<ITextModel>();
     readonly onDidSaveModel = this.onDidSaveModelEmitter.event;
 
-    protected readonly onWillSaveModelEmitter = new Emitter<WillSaveMonacoModelEvent>();
-    readonly onWillSaveModel = this.onWillSaveModelEmitter.event;
-
     protected readonly onDidChangeValidEmitter = new Emitter<void>();
     readonly onDidChangeValid = this.onDidChangeValidEmitter.event;
 
@@ -99,6 +91,8 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
 
     protected resourceVersion: ResourceVersion | undefined;
 
+    protected readonly willSaveModelListeners: WillSaveMonacoModelListener[] = [];
+
     constructor(
         protected readonly resource: Resource,
         protected readonly m2p: MonacoToProtocolConverter,
@@ -110,7 +104,6 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
         this.toDispose.push(this.toDisposeOnAutoSave);
         this.toDispose.push(this.onDidChangeContentEmitter);
         this.toDispose.push(this.onDidSaveModelEmitter);
-        this.toDispose.push(this.onWillSaveModelEmitter);
         this.toDispose.push(this.onDirtyChangedEmitter);
         this.toDispose.push(this.onDidChangeValidEmitter);
         this.toDispose.push(Disposable.create(() => this.cancelSave()));
@@ -154,7 +147,7 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
         if (mode === EncodingMode.Decode) {
             return this.sync();
         }
-        return this.scheduleSave(TextDocumentSaveReason.Manual, this.cancelSave(), true);
+        return this.scheduleSave(this.cancelSave(), true, { saveReason: SaveReason.Manual });
     }
 
     getEncoding(): string | undefined {
@@ -386,7 +379,10 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
     }
 
     save(options?: SaveOptions): Promise<void> {
-        return this.scheduleSave(options?.saveReason ?? TextDocumentSaveReason.Manual, undefined, undefined, options);
+        return this.scheduleSave(undefined, undefined, {
+            saveReason: TextDocumentSaveReason.Manual,
+            ...options
+        });
     }
 
     protected pendingOperation = Promise.resolve();
@@ -485,8 +481,8 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
         return this.saveCancellationTokenSource.token;
     }
 
-    protected scheduleSave(reason: TextDocumentSaveReason, token: CancellationToken = this.cancelSave(), overwriteEncoding?: boolean, options?: SaveOptions): Promise<void> {
-        return this.run(() => this.doSave(reason, token, overwriteEncoding, options));
+    protected scheduleSave(token: CancellationToken = this.cancelSave(), overwriteEncoding?: boolean, options?: SaveOptions): Promise<void> {
+        return this.run(() => this.doSave(token, overwriteEncoding, options));
     }
 
     protected ignoreContentChanges = false;
@@ -546,18 +542,18 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
         }
     }
 
-    protected async doSave(reason: TextDocumentSaveReason, token: CancellationToken, overwriteEncoding?: boolean, options?: SaveOptions): Promise<void> {
+    protected async doSave(token: CancellationToken, overwriteEncoding?: boolean, options?: SaveOptions): Promise<void> {
         if (token.isCancellationRequested || !this.resource.saveContents) {
             return;
         }
 
-        await this.fireWillSaveModel(reason, token, options);
+        await this.fireWillSaveModel(token, options);
         if (token.isCancellationRequested) {
             return;
         }
 
         const changes = [...this.contentChanges];
-        if ((changes.length === 0 && !this.resource.initiallyDirty) && !overwriteEncoding && reason !== TextDocumentSaveReason.Manual) {
+        if ((changes.length === 0 && !this.resource.initiallyDirty) && !overwriteEncoding && options?.saveReason !== TextDocumentSaveReason.Manual) {
             return;
         }
 
@@ -586,64 +582,19 @@ export class MonacoEditorModel implements IResolvedTextEditorModel, TextEditorDo
         }
     }
 
-    protected async fireWillSaveModel(reason: TextDocumentSaveReason, token: CancellationToken, options?: SaveOptions): Promise<void> {
-        type EditContributor = Thenable<monaco.editor.IIdentifiedSingleEditOperation[]>;
-
-        const firing = this.onWillSaveModelEmitter.sequence(async listener => {
-            if (token.isCancellationRequested) {
-                return false;
+    registerWillSaveModelListener(listener: WillSaveMonacoModelListener): Disposable {
+        this.willSaveModelListeners.push(listener);
+        return Disposable.create(() => {
+            const index = this.willSaveModelListeners.indexOf(listener);
+            if (index >= 0) {
+                this.willSaveModelListeners.splice(index, 1);
             }
-            const waitables: EditContributor[] = [];
-            const { version } = this;
-
-            const event = {
-                model: this, reason, options,
-                waitUntil: (thenable: EditContributor) => {
-                    if (Object.isFrozen(waitables)) {
-                        throw new Error('waitUntil cannot be called asynchronously.');
-                    }
-                    waitables.push(thenable);
-                }
-            };
-
-            // Fire.
-            try {
-                listener(event);
-            } catch (err) {
-                console.error(err);
-                return true;
-            }
-
-            // Asynchronous calls to `waitUntil` should fail.
-            Object.freeze(waitables);
-
-            // Wait for all promises.
-            const edits = await Promise.all(waitables).then(allOperations =>
-                ([] as monaco.editor.IIdentifiedSingleEditOperation[]).concat(...allOperations)
-            );
-            if (token.isCancellationRequested) {
-                return false;
-            }
-
-            // In a perfect world, we should only apply edits if document is clean.
-            if (version !== this.version) {
-                console.error('onWillSave listeners should provide edits, not directly alter the document.');
-            }
-
-            // Finally apply edits provided by this listener before firing the next.
-            if (edits && edits.length > 0) {
-                this.applyEdits(edits, {
-                    ignoreDirty: true,
-                });
-            }
-
-            return true;
         });
+    }
 
-        try {
-            await firing;
-        } catch (e) {
-            console.error(e);
+    protected async fireWillSaveModel(token: CancellationToken, options?: SaveOptions): Promise<void> {
+        for (const listener of this.willSaveModelListeners) {
+            await listener(this, token, options);
         }
     }